feat(files): Display mtime in grid view
